+++ source/builder.hpp
@@ -93,16 +93,16 @@ namespace hummus {
                         auto& dendriticSynapse = input.back()["neurons"][n->get_neuron_id()]["dendritic_synapses"];
                         if (dendriticSynapse.is_array() && !dendriticSynapse.empty()) {
                             for (auto i=0; i<dendriticSynapse.size(); i++) {
-                                double weight = 0;
+                                float weight = 0;
                                 if (dendriticSynapse[i]["weight"].is_number()) {
-                                    weight = dendriticSynapse[i]["weight"].get<double>();
+                                    weight = dendriticSynapse[i]["weight"].get<float>();
                                 } else {
                                     throw std::logic_error("dendritic synapse weight incorrectly formatted");
                                 }
 
-                                double delay = 0;
+                                float delay = 0;
                                 if (dendriticSynapse[i]["delay"].is_number()) {
-                                    delay = dendriticSynapse[i]["delay"].get<double>();
+                                    delay = dendriticSynapse[i]["delay"].get<float>();
                                 } else {
                                     throw std::logic_error("dendritic synapse weight incorrectly formatted");
                                 }
@@ -137,15 +137,15 @@ namespace hummus {
             }
             
             if (input["trace_time_constant"].is_number()) {
-                n->set_trace_time_constant(input["trace_time_constant"].get<double>());
+                n->set_trace_time_constant(input["trace_time_constant"].get<float>());
             }
             
             if (input["resting_potential"].is_number()) {
-                n->set_resting_potential(input["resting_potential"].get<double>());
+                n->set_resting_potential(input["resting_potential"].get<float>());
             }
             
             if (input["threshold"].is_number()) {
-                n->set_threshold(input["threshold"].get<double>());
+                n->set_threshold(input["threshold"].get<float>());
             }
             
             if (input["refractory_period"].is_number()) {
@@ -153,15 +153,15 @@ namespace hummus {
             }
             
             if (input["membrane_time_constant"].is_number()) {
-                n->set_membrane_time_constant(input["membrane_time_constant"].get<double>());
+                n->set_membrane_time_constant(input["membrane_time_constant"].get<float>());
             }
             
             if (input["conductance"].is_number()) {
-                n->set_conductance(input["conductance"].get<double>());
+                n->set_conductance(input["conductance"].get<float>());
             }
             
             if (input["leakage_conductance"].is_number()) {
-                n->set_leakage_conductance(input["leakage_conductance"].get<double>());
+                n->set_leakage_conductance(input["leakage_conductance"].get<float>());
             }
             
             if (input["class_label"].is_string()) {
@@ -185,29 +185,29 @@ namespace hummus {
             if (axonalSynapse.is_array() && !axonalSynapse.empty()) {
 
                 for (auto i=0; i<axonalSynapse.size(); i++) {
-                    double weight = 0;
+                    float weight = 0;
                     if (axonalSynapse[i]["weight"].is_number()) {
-                        weight = axonalSynapse[i]["weight"].get<double>();
+                        weight = axonalSynapse[i]["weight"].get<float>();
                     } else {
                         throw std::logic_error("axonal synapse weight incorrectly formatted");
                     }
 
-                    double delay = 0;
+                    float delay = 0;
                     if (axonalSynapse[i]["delay"].is_number()) {
-                        delay = axonalSynapse[i]["delay"].get<double>();
+                        delay = axonalSynapse[i]["delay"].get<float>();
                     } else {
                         throw std::logic_error("axonal synapse weight incorrectly formatted");
                     }
                     
                     if (axonalSynapse[i]["postsynaptic_neuron"].is_number()) {
-                        double synapseTimeConstant = 0;
+                        float synapseTimeConstant = 0;
                         int json_id = axonalSynapse[i]["json_id"].get<int>();
                         
                         switch (json_id) {
                             case 0: {
-                                double amplitudeScaling = 0;
+                                float amplitudeScaling = 0;
                                 if (axonalSynapse[i]["amplitude_scaling"].is_number()) {
-                                    amplitudeScaling = axonalSynapse[i]["amplitude_scaling"].get<double>();
+                                    amplitudeScaling = axonalSynapse[i]["amplitude_scaling"].get<float>();
                                 } else {
                                     throw std::logic_error("dirac synapse amplitude scaling incorrectly formatted");
                                 }
@@ -217,7 +217,7 @@ namespace hummus {
                                 break;
                             } case 1: {
                                 if (axonalSynapse[i]["synapse_time_constant"].is_number()) {
-                                    synapseTimeConstant = axonalSynapse[i]["synapse_time_constant"].get<double>();
+                                    synapseTimeConstant = axonalSynapse[i]["synapse_time_constant"].get<float>();
                                 } else {
                                     throw std::logic_error("exponential synaptic time constant incorrectly formatted");
                                 }
@@ -227,7 +227,7 @@ namespace hummus {
                                 break;
                             } case 2:
                                 if (axonalSynapse[i]["synapse_time_constant"].is_number()) {
-                                    synapseTimeConstant = axonalSynapse[i]["synapse_time_constant"].get<double>();
+                                    synapseTimeConstant = axonalSynapse[i]["synapse_time_constant"].get<float>();
                                 } else {
                                     throw std::logic_error("pulse synaptic time constant incorrectly formatted");
                                 }
@@ -283,7 +283,7 @@ namespace hummus {
             }
 
             if (input["decay_homeostasis"].is_number()) {
-                dynamic_cast<T*>(n)->set_decay_homeostasis(input["decay_homeostasis"].get<double>());
+                dynamic_cast<T*>(n)->set_decay_homeostasis(input["decay_homeostasis"].get<float>());
             }
 
             if (input["homeostasis"].is_boolean()) {
@@ -291,11 +291,11 @@ namespace hummus {
             }
             
             if (input["homeostasis_beta"].is_number()) {
-                dynamic_cast<T*>(n)->set_homeostasis_beta(input["homeostasis_beta"].get<double>());
+                dynamic_cast<T*>(n)->set_homeostasis_beta(input["homeostasis_beta"].get<float>());
             }
 
             if (input["resting_threshold"].is_number()) {
-                dynamic_cast<T*>(n)->set_resting_threshold(input["resting_threshold"].get<double>());
+                dynamic_cast<T*>(n)->set_resting_threshold(input["resting_threshold"].get<float>());
             }
         }
